Added some apt packages

+++ bash_aliases
@@ -50,8 +50,9 @@ doPsKill(){
 }
 alias pskill=doPsKill
 
-alias docker-nuke="docker rm $(docker ps -a -q) && docker rmi $(docker images -q)"
-alias dc='docker-compose'
-alias docker-rm-untagged='docker rmi -f $(docker images | grep "^<none>" | awk "{print $3}")'
-
+if ! type "docker" >  /dev/null; then
+    alias docker-nuke="docker rm $(docker ps -a -q) && docker rmi $(docker images -q)"
+    alias dc='docker-compose'
+    alias docker-rm-untagged='docker rmi -f $(docker images | grep "^<none>" | awk "{print $3}")'
+fi
